make FieldMinMaxValueJsonConverter internal and move to own file
